Mainly making all cogs inherit from GuildedCog but also moving some stuff round
